requirements.txt
python-dotenv==1.0.1
aiogram==3.18.0
asyncpg==0.30.0
###
###